# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Battleship API",
    "description": "REST API for the Battleship game backend (session, preparation, gameplay)",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/v2/game/sessions": {
      "post": {
        "tags": [
          "Game Session"
        ],
        "summary": "Create a game session",
        "description": "Creates a new session for the given game edition. Fails if the edition name is blank or not a recognized GameEdition.",
        "operationId": "createGameSession",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParamGameEditionDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Session created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseCreatedSessionIdDto"
                },
                "examples": {
                  "success": {
                    "description": "success",
                    "value": {
                      "sessionId": "abc123"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid or unrecognized game edition",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionDto"
                },
                "examples": {
                  "editionInvalid": {
                    "description": "editionInvalid",
                    "value": {
                      "status": 400,
                      "errorMessage": "...",
                      "errorCode": "EDITION_INVALID"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/game/sessions/{sessionId}/players": {
      "post": {
        "tags": [
          "Game Session"
        ],
        "summary": "Create a player in a session",
        "description": "Registers a new player under the given session. Fails if the player name is blank, the session id is unknown, the session already has 2 players, or the session is not in a stage that allows joining.",
        "operationId": "createPlayerInSession",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "description": "Session identifier returned by session creation",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "abc123"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParamPlayerNameDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Player created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseCreatedPlayerDto"
                },
                "examples": {
                  "success": {
                    "description": "success",
                    "value": {
                      "playerId": "p1",
                      "playerName": "Alice"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid player name, unknown session, session full, or wrong stage",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionDto"
                },
                "examples": {
                  "playerNameInvalid": {
                    "description": "playerNameInvalid",
                    "value": {
                      "status": 400,
                      "errorMessage": "...",
                      "errorCode": "PLAYER_NAME_INVALID"
                    }
                  },
                  "sessionNotFound": {
                    "description": "sessionNotFound",
                    "value": {
                      "status": 400,
                      "errorMessage": "...",
                      "errorCode": "SESSION_NOT_FOUND"
                    }
                  },
                  "sessionFull": {
                    "description": "sessionFull",
                    "value": {
                      "status": 400,
                      "errorMessage": "...",
                      "errorCode": "SESSION_FULL"
                    }
                  },
                  "stageInvalid": {
                    "description": "stageInvalid",
                    "value": {
                      "status": 400,
                      "errorMessage": "...",
                      "errorCode": "STAGE_INVALID"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Unexpected internal error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionDto"
                },
                "examples": {
                  "internal": {
                    "description": "internal",
                    "value": {
                      "status": 500,
                      "errorMessage": "...",
                      "errorCode": "INTERNAL"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/game/sessions/{sessionId}/state": {
      "get": {
        "tags": [
          "Game Session"
        ],
        "summary": "Get current game stage",
        "description": "Returns the session's current GameStage. Fails if the session id is unknown.",
        "operationId": "getCurrentGameStage",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "description": "Session identifier",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "abc123"
          }
        ],
        "responses": {
          "200": {
            "description": "Current stage",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseCurrentGameStageDto"
                },
                "examples": {
                  "success": {
                    "description": "success",
                    "value": {
                      "gameStage": "PREPARATION"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Unknown session id",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionDto"
                },
                "examples": {
                  "sessionNotFound": {
                    "description": "sessionNotFound",
                    "value": {
                      "status": 400,
                      "errorMessage": "...",
                      "errorCode": "SESSION_NOT_FOUND"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/game/sessions/{sessionId}/changesTime": {
      "get": {
        "tags": [
          "Game Session"
        ],
        "summary": "Get last session change timestamp",
        "description": "Returns the timestamp of the session's last state change, used by clients for change-detection polling. Fails if the session id is unknown.",
        "operationId": "getLastSessionChangeTime",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "description": "Session identifier",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "abc123"
          }
        ],
        "responses": {
          "200": {
            "description": "Last change timestamp",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseLastSessionChangeTimeDto"
                },
                "examples": {
                  "success": {
                    "description": "success",
                    "value": {
                      "lastId": "2026-07-13T10:15:30"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Unknown session id",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionDto"
                },
                "examples": {
                  "sessionNotFound": {
                    "description": "sessionNotFound",
                    "value": {
                      "status": 400,
                      "errorMessage": "...",
                      "errorCode": "SESSION_NOT_FOUND"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExceptionDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP status code",
            "example": 400
          },
          "errorMessage": {
            "type": "string",
            "description": "Human-readable error message"
          },
          "errorCode": {
            "type": "string",
            "description": "Stable, machine-readable error code identifying the exception type",
            "example": "SESSION_NOT_FOUND"
          }
        }
      },
      "ParamGameEditionDto": {
        "type": "object",
        "properties": {
          "gameEdition": {
            "type": "string",
            "description": "Game edition name",
            "enum": [
              "UKRAINIAN",
              "MILTON_BRADLEY"
            ]
          }
        }
      },
      "ResponseCreatedSessionIdDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "description": "Newly created session's identifier"
          }
        }
      },
      "ParamPlayerNameDto": {
        "type": "object",
        "properties": {
          "playerName": {
            "type": "string",
            "description": "Display name for the new player"
          }
        }
      },
      "ResponseCreatedPlayerDto": {
        "type": "object",
        "properties": {
          "playerId": {
            "type": "string",
            "description": "Newly created player's identifier"
          },
          "playerName": {
            "type": "string",
            "description": "Player's display name"
          }
        }
      },
...
      "ResponseCurrentGameStageDto": {
        "type": "object",
        "properties": {
          "gameStage": {
            "type": "string",
            "description": "Current GameStage name",
            "enum": [
              "INITIALIZED",
              "WAITING_FOR_PLAYERS",
              "PREPARATION",
              "IN_GAME",
              "FINISHED"
            ]
          }
        }
      },
...
      "ResponseLastSessionChangeTimeDto": {
        "type": "object",
        "properties": {
          "lastId": {
            "type": "string",
            "description": "Last-update timestamp as a string, used for client polling comparisons"
          }
        }
      },
...
    }
  }
}